[#28] refact(views) - 폴더 정리 및 뷰 이름에서 Invoice 제거
