+++ frontend/view/RouterView/ChartView.ts
@@ -1,6 +1,6 @@
-import { View } from './index'
+import { View } from '../index'
 
-export default class InvoiceChartView extends View {
+export default class ChartView extends View {
   constructor() {
     super('invoice-chart', 'section')
   }
